asynMotorController: Handle CommsError in readGenericPointer()
readGenericPointer() does check that the driver has but at least
one bit into the status bits.
Add a check that the CommsError bit is not set:
It means (or could mean) that there had been no communication
to the controller yet.
As a consequence, the current motorRecord stays in UDF alarm.