motorRecord: Do not do retries after slipstall
When a motor reports a "slip/stall" situation, do not try to retry.
Having msta.Bits.EA_SLIP_STALL set means that there is a mechanical problem
that needs to investigated and fixed.
The controller/driver is responsible to stop here,
and the motorRecord should not command an "automatic retry" -
so don't do that any more