motorRecord: another update after LS error: Endless loop
Working with a setup that has the .DLY fiels set to a value > 0.0
showed a bug when the motor run into a limit switch:
The statmachine run into an endless loop.
The code around "Do another update after LS error"
was triggered again and again.
Problem:
After the delay had elapsed, the MIP field was not MIP_DONE but
DELAY_REQUEST, and that triggered and re-triggered the
call to devSupGetInfo() again and again.

Solution:
When delay is used, skip over the call to devSupGetInfo() done
"right now".
This is OK since the delay handling will trigger a devSupGetInfo()
later.