asynMotorController.cpp: Avoid useless power on
Improve the situation when "auto power on" is used:
When the motor is already powered, there is no need to calll
setClosedLoop(1) at all.

Note:
setClosedLoop() is used to control the power to the motor,
enable or disable it.
This is because all model 3 drivers use this
method and closed loop is handled inside the controller.
And now it is too late to change it.